requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
python-dotenv==0.19.0
aiofiles==23.2.1
jinja2==3.0.3